refactor: reorg templates
